move validation related code to separate file in utils
